deployment/application_deployment/requirements.txt
aws-cdk.core==1.92.0
aws-cdk.aws-ec2==1.92.0
aws-cdk.aws-iam==1.92.0
aws-cdk.aws-secretsmanager==1.92.0
